Migrate Serialization store to its own module
